add test to check healthpoint
