+++ .github/workflows/pipeline.yaml
@@ -21,5 +21,7 @@ jobs:
         run: npm run build
       - name: test
         run: npm run test
+      - name: run
+        run: npm run start-prod &&
       - name: Trigger deployment
         run: curl https://api.render.com/deploy/srv-${{ secrets.RENDER_SERVICE_ID }}?key=${{ secrets.RENDER_API_KEY }}
